+++ src/get/_get.js
@@ -2,6 +2,7 @@
 // exporting modules to be included the UMD bundle
 
 import activeElements from './active-elements';
+import focusRedirectTarget from './focus-redirect-target';
 import focusTarget from './focus-target';
 import insignificantBranches from './insignificant-branches';
 import parents from './parents';
@@ -9,6 +10,7 @@ import shadowHostParents from './shadow-host-parents';
 import shadowHost from './shadow-host';
 export default {
   activeElements,
+  focusRedirectTarget,
   focusTarget,
   insignificantBranches,
   parents,